Factor out a dwh.util module.
